+++ packages/i18n/src/Loader.js
@@ -1,4 +1,4 @@
-/* global XMLHttpRequest, ILIB_BASE_PATH, ILIB_RESOURCES_PATH, ILIB_CACHE_ID */
+/* global XMLHttpRequest, ILIB_BASE_PATH, ILIB_RESOURCES_PATH, ILIB_ADDITIONAL_RESOURCES_PATH, ILIB_CACHE_ID */
 
 import {memoize} from '@enact/core/util';
 import Loader from 'ilib/lib/Loader';
@@ -51,7 +51,6 @@ const cachePrefix = 'ENACT-ILIB-';
 const cacheKey = cachePrefix + 'CACHE-ID';
 const cacheID = typeof ILIB_CACHE_ID === 'undefined' ? '$ILIB' : ILIB_CACHE_ID;
 const timeStampKey = 'l10n_timestamp';
-const iLibAdditionalPath = process.env.REACT_APP_ILIB_ADDITIONAL_RESOURCES_PATH;
 
 function EnyoLoader () {
 	this.base = iLibBase;
@@ -244,20 +243,27 @@ EnyoLoader.prototype.loadFiles = function (paths, sync, params, callback, rootPa
 				};
 
 				const handleAdditionalResourcesPath = (json, err) => {
-					if (found && !err && typeof json === 'object') {
-						// Overwrite the _root/path result
-						Object.assign(ret[ret.length - 1], json);
+					if (!err && typeof json === 'object') {
+						if (found) {
+							// Overwrite the _root/path result
+							Object.assign(ret[ret.length - 1], json);
+						} else {
+							// This case is where the file is only in the additional resources path
+							cache.update = true;
+							ret.push(json);
+							found = true;
+						}
 					}
 				};
 
 				if (this.isAvailable(_root, path)) {
 					getSync(this._pathjoin(_root, path), handler);
+				}
 
-					if (this.addPaths && Array.isArray(this.addPaths)) {
-						for (let addedRoot of this.addPaths) {
-							if (this.isAvailable(addedRoot, path)) {
-								getSync(this._pathjoin(addedRoot, path), handleAdditionalResourcesPath);
-							}
+				if (this.addPaths && Array.isArray(this.addPaths)) {
+					for (let addedRoot of this.addPaths) {
+						if (this.isAvailable(addedRoot, path)) {
+							getSync(this._pathjoin(addedRoot, path), handleAdditionalResourcesPath);
 						}
 					}
 				}
@@ -290,7 +296,7 @@ EnyoLoader.prototype.loadFiles = function (paths, sync, params, callback, rootPa
 		// asynchronous
 		let cache = {data: this._loadFilesCache(_root, paths)};
 
-		Promise.all(paths.map(path => this._loadFilesAsync(path, params, cache))).then(results => {
+		Promise.all(paths.map(path => this._loadFilesAsync(path, params, cache, rootPath))).then(results => {
 			if (cache.update) {
 				this._storeFilesCache(_root, paths, results);
 			}
@@ -302,7 +308,7 @@ EnyoLoader.prototype.loadFiles = function (paths, sync, params, callback, rootPa
 };
 
 EnyoLoader.prototype._handleManifest = function (dirpath, filepath, json) {
-	const isAdditionalPath = typeof iLibAdditionalPath !== 'undefined' ? dirpath.includes(iLibAdditionalPath) : false;
+	const isAdditionalPath = typeof ILIB_ADDITIONAL_RESOURCES_PATH !== 'undefined' ? dirpath.includes(ILIB_ADDITIONAL_RESOURCES_PATH) : false;
 	// star indicates there was no ilibmanifest.json, so always try to load files from
 	// that dir
 	if (json != null) {
@@ -342,7 +348,7 @@ EnyoLoader.prototype._validateManifest = function (cachedManifest, filepath, syn
 					newManifest = json;
 				});
 			}
-			if (newManifest === null && typeof iLibAdditionalPath !== 'undefined' && filepath.includes(iLibAdditionalPath)) {
+			if (newManifest === null && typeof ILIB_ADDITIONAL_RESOURCES_PATH !== 'undefined' && filepath.includes(ILIB_ADDITIONAL_RESOURCES_PATH)) {
 				// If new manifest is null and the filepath has ILIB_ADDITIONAL_RESOURCES_PATH,
 				// meaning we need to clear string cache
 				this._clearStringsCache();