+++ packages/ui/Layout/Layout.js
@@ -6,8 +6,8 @@
  * which were largely abandoned for their drawbacks, ignoring their strengths. A `Layout` is simply
  * a container for `Cell`, the only "legal" child. Conversely, `Cell` may only be used in a
  * `Layout`. Cells in a Layout can either be positioned next to each other (horizontally) or
- * above/below each other (vertically) in what we refer to as a [Row]{@link ui/Layout.Row} or
- * [Column]{@link ui/Layout.Column}, respectively.
+ * above/below each other (vertically) in what we refer to as a {@link ui/Layout.Row|Row} or
+ * {@link ui/Layout.Column|Column}, respectively.
  *
  * The `Row` and `Column` layout presets describe the direction of layout for their children. This
  * can sometimes cause confusion. A `Row` of children naturally forms a _layout_ whose children can
@@ -136,9 +136,9 @@ import css from './Layout.module.less';
 /**
  * A container for `Cell`s.
  *
- * A stateless component that acts as a containing area for [Cells]{@link ui/Layout.Cell} to be
+ * A stateless component that acts as a containing area for {@link ui/Layout.Cell|Cells} to be
  * positioned in a row or a column (horizontally or vertically, respectively. It supports an
- * [orientation]{@link ui/Layout.Layout#orientation} property for laying-out its contents
+ * {@link ui/Layout.LayoutBase.orientation|orientation} property for laying-out its contents
  * (`Cells`) in an organized, readable way.
  *
  * Example:
@@ -170,16 +170,37 @@ const LayoutBase = kind({
 		/**
 		 * The alignment of children.
 		 *
-		 * Aligns the children [Cells]{@link ui/Layout.Cell} vertically in the case of a horizontal
+		 * Aligns the children {@link ui/Layout.Cell|Cells} vertically in the case of a horizontal
 		 * layout or horizontally in the case of a vertical layout. `"start"`, `"center"` and
 		 * `"end"` are the most commonly used, although all values of `align-items` are supported.
 		 * `"start"` refers to the top in a horizontal layout, and left in a vertical LTR layout
 		 * `"end"` refers to the bottom in a horizontal layout, and right in a vertical LTR layout
 		 * `"start"` and `"end"` reverse places when in a vertical layout in a RTL locale.
-		 * This includes support for `align-parts` which is shorthand for combining `align-items`
-		 * and `justify-content` into a single property, separated by a space, in that order.
-		 * This allows you to specify both the horizontal and vertical alignment in one property,
-		 * separated by a space.
+		 * This includes support shorthand for combining `align-items` and `justify-content` into
+		 * a single property, separated by a space, in that order. This allows you to specify both
+		 * the horizontal and vertical alignment in one property, separated by a space.
+		 *
+		 * For example, `align="center space-between"` means `align-items: center` and
+		 * `justify-content: space-between` for each. `justify-content` property can be used to
+		 * align the Cells on the main axis and adjust gaps among the Cells. To declare the
+		 * `justify-content` property only, just add a heading space for `align` prop string like
+		 * `align=" space-between"`. The default value for `align-items` is `stretch`. It can be
+		 * also used for `align` prop like `align="stretch space-between"`. All values of
+		 * `justify-content` are supported, like `start`, `end`, `center`, `space-between`,
+		 * `space-around`, and `space-evenly`.
+		 *
+		 * ```
+		 * <Layout align="center space-around">
+		 *     <Cell>Left Column</Cell>
+		 *     <Cell>Right Column</Cell>
+		 * </Layout>
+		 * ```
+		 * ```
+		 * <Layout align=" space-between">
+		 *     <Cell>Left Column</Cell>
+		 *     <Cell>Right Column</Cell>
+		 * </Layout>
+		 * ```
 		 *
 		 * @type {String}
 		 * @public
@@ -187,7 +208,7 @@ const LayoutBase = kind({
 		align: PropTypes.string,
 
 		/**
-		 * Only [Cell]{@link ui/Layout.Cell} components are supported as children.
+		 * Only {@link ui/Layout.Cell|Cell} components are supported as children.
 		 *
 		 * @type {Cell|Cell[]}
 		 * @public
@@ -226,7 +247,7 @@ const LayoutBase = kind({
 		inline: PropTypes.bool,
 
 		/**
-		 * The orientation of the `Layout`, i.e. how the children [Cells]{@link ui/Layout.Cell} are
+		 * The orientation of the `Layout`, i.e. how the children {@link ui/Layout.Cell|Cells} are
 		 * positioned on the screen. Must be either `'horizontal'` or `'vertical'`.
 		 *
 		 * @type {String}
@@ -312,9 +333,9 @@ const LayoutDecorator = ForwardRef({prop: 'componentRef'});
 /**
  * A container for `Cell`s.
  *
- * A stateless component that acts as a containing area for [Cells]{@link ui/Layout.Cell} to be
+ * A stateless component that acts as a containing area for {@link ui/Layout.Cell|Cells} to be
  * positioned in a row or a column (horizontally or vertically, respectively. It supports an
- * [orientation]{@link ui/Layout.Layout#orientation} property for laying-out its contents
+ * {@link ui/Layout.LayoutBase.orientation|orientation} property for laying-out its contents
  * (`Cells`) in an organized, readable way.
  *
  * Example:
@@ -346,7 +367,7 @@ const Layout = LayoutDecorator(LayoutBase);
 
 /**
  * Shorthand for `<Layout orientation="vertical">`, which positions its
- * [Cells]{@link ui/Layout.Cell} vertically.
+ * {@link ui/Layout.Cell|Cells} vertically.
  * ```
  * ┌────┐
  * ├────┤
@@ -372,7 +393,7 @@ Column.displayName = 'Column';
 
 /**
  * Shorthand for `<Layout orientation="horizontal">`, which positions its
- * [Cells]{@link ui/Layout.Cell} horizontally.
+ * {@link ui/Layout.Cell|Cells} horizontally.
  * ```
  * ┌─┬─┬─┬─┐
  * │ │ │ │ │