+++ packages/ui/ToggleItem/ToggleItem.js
@@ -3,14 +3,16 @@
  *
  * This is not intended to be used directly, but should be extended by a component that will
  * customize this component's appearance by supplying an
- * [iconComponent prop]{@link ui/ToggleItem.ToggleItemBase#iconComponent}.
+ * {@link ui/ToggleItem.ToggleItemBase.iconComponent|iconComponent} prop.
  *
  * @module ui/ToggleItem
  * @exports ToggleItem
  * @exports ToggleItemBase
  * @exports ToggleItemDecorator
+ * @deprecated Will be removed in 5.0.0.
  */
 
+import deprecate from '@enact/core/internal/deprecate';
 import EnactPropTypes from '@enact/core/internal/prop-types';
 import kind from '@enact/core/kind';
 import PropTypes from 'prop-types';
@@ -88,7 +90,7 @@ const ToggleItemBase = kind({
 		 *
 		 * This receives the `css` prop for theme extension and therefore must be a custom
 		 * component and not a simple HTML DOM node. Recommended component or themed
-		 * derivative: [SlotItem]{@link ui/SlotItem.SlotItem}
+		 * derivative: {@link ui/SlotItem.SlotItem|SlotItem}
 		 *
 		 * @type {Component}
 		 * @required
@@ -101,7 +103,7 @@ const ToggleItemBase = kind({
 		 *
 		 * This component receives the `selected` prop and value,
 		 * and must therefore respond to it in some way. It is recommended to use
-		 * [ToggleIcon]{@link ui/ToggleIcon} for this.
+		 * {@link ui/ToggleIcon|ToggleIcon} for this.
 		 *
 		 * @type {Component|Element}
 		 * @required
@@ -145,7 +147,7 @@ const ToggleItemBase = kind({
 		/**
 		 * An optional prop that lets you override the icon of the `iconComponent` component.
 		 *
-		 * This accepts any string that the [Icon]{@link ui/Icon.Icon} component supports, provided
+		 * This accepts any string that the {@link ui/Icon.Icon|Icon} component supports, provided
 		 * the recommendations of `iconComponent` are followed.
 		 *
 		 * @type {String|Object}
@@ -247,7 +249,7 @@ const ToggleItemBase = kind({
 		slotAfter: iconCreator('after')
 	},
 
-	render: ({component: Component, componentRef, css, children, selected, ...rest}) => {
+	render: deprecate(({component: Component, componentRef, css, children, selected, ...rest}) => {
 		delete rest.iconComponent;
 		delete rest.iconPosition;
 		delete rest.itemIcon;
@@ -265,7 +267,10 @@ const ToggleItemBase = kind({
 				{children}
 			</Component>
 		);
-	}
+	}, {
+		name: 'ui/ToggleItem',
+		until: '5.0.0'
+	})
 });
 
 /**