+++ packages/ui/VirtualList/UiVirtualListBase.js
@@ -1,6 +1,6 @@
-/* eslint-disable react/no-this-in-sfc */
 import classNames from 'classnames';
 import {forward} from '@enact/core/handle';
+import deprecate from '@enact/core/internal/deprecate';
 import {platform} from '@enact/core/platform';
 import PropTypes from 'prop-types';
 import equals from 'ramda/src/equals';
@@ -83,7 +83,7 @@ const VirtualListBaseFactory = (type) => {
 
 			/**
 			 * Callback method of scrollTo.
-			 * Normally, [Scrollable]{@link ui/Scrollable.Scrollable} should set this value.
+			 * Normally, {@link ui/Scrollable.Scrollable|Scrollable} should set this value.
 			 *
 			 * @type {Function}
 			 * @private
@@ -1122,6 +1122,10 @@ const VirtualListBaseFactory = (type) => {
 		}
 
 		render () {
+			deprecate({
+				name: 'ui/VirtualList.VirtualListBase',
+				until: '5.0.0'
+			});
 			const
 				{className, containerProps, itemsRenderer, style, ...rest} = this.props,
 				{cc, itemContainerRef, primary} = this,
@@ -1164,19 +1168,20 @@ const VirtualListBaseFactory = (type) => {
 
 /**
  * A basic base component for
- * [VirtualList]{@link ui/VirtualList.VirtualList} and [VirtualGridList]{@link ui/VirtualList.VirtualGridList}.
+ * {@link ui/VirtualList.VirtualList|VirtualList} and {@link ui/VirtualList.VirtualGridList|VirtualGridList}.
  *
  * @class VirtualListBase
  * @memberof ui/VirtualList
  * @ui
  * @public
+ * @deprecated Will be removed in 5.0.0.
  */
 const VirtualListBase = VirtualListBaseFactory(JS);
 VirtualListBase.displayName = 'ui:VirtualListBase';
 
 /**
  * A basic base component for
- * [VirtualListNative]{@link ui/VirtualList.VirtualListNative} and [VirtualGridListNative]{@link ui/VirtualList.VirtualGridListNative}.
+ * {@link ui/VirtualList.VirtualListNative|VirtualListNative} and {@link ui/VirtualList.VirtualGridListNative|VirtualGridListNative}.
  *
  * @class VirtualListBaseNative
  * @memberof ui/VirtualList