+++ packages/ui/useScroll/useScroll.js
@@ -162,6 +162,7 @@ const useScrollBase = (props) => {
 		// status
 		deferScrollTo: true,
 		isScrollAnimationTargetAccumulated: false,
+		rtl,
 
 		// overscroll
 		lastInputType: null,
@@ -214,74 +215,88 @@ const useScrollBase = (props) => {
 		prevState: {isHorizontalScrollbarVisible, isVerticalScrollbarVisible}
 	});
 
+	/* istanbul ignore next */
+	const themeScrollContainerHandle = useRef({
+		get animator () {
+			return mutableRef.current.animator;
+		},
+		get bounds () {
+			return mutableRef.current.bounds;
+		},
+		get isDragging () {
+			return mutableRef.current.isDragging;
+		},
+		set isDragging (val) {
+			mutableRef.current.isDragging = val;
+		},
+		get isScrollAnimationTargetAccumulated () {
+			return mutableRef.current.isScrollAnimationTargetAccumulated;
+		},
+		set isScrollAnimationTargetAccumulated (val) {
+			mutableRef.current.isScrollAnimationTargetAccumulated = val;
+		},
+		get lastInputType () {
+			return mutableRef.current.lastInputType;
+		},
+		set lastInputType (val) {
+			mutableRef.current.lastInputType = val;
+		},
+		get rtl () {
+			return mutableRef.current.rtl;
+		},
+		get scrollBounds () {
+			return getScrollBounds();
+		},
+		get scrollHeight () {
+			return mutableRef.current.bounds.scrollHeight;
+		},
+		get scrolling () {
+			return mutableRef.current.scrolling;
+		},
+		get scrollLeft () {
+			return mutableRef.current.scrollLeft;
+		},
+		get scrollToInfo () {
+			return mutableRef.current.scrollToInfo;
+		},
+		get scrollTop () {
+			return mutableRef.current.scrollTop;
+		},
+		get wheelDirection () {
+			return mutableRef.current.wheelDirection;
+		},
+		set wheelDirection (val) {
+			mutableRef.current.wheelDirection = val;
+		}
+	});
+
 	if (mutableRef.current.animator == null) {
 		mutableRef.current.animator = new ScrollAnimator();
 	}
 
+	if (mutableRef.current.rtl !== rtl) {
+		mutableRef.current.rtl = rtl;
+	}
+
 	useLayoutEffect(() => {
+		/* istanbul ignore next */
 		if (setScrollContainerHandle) {
-			setScrollContainerHandle({
-				animator: mutableRef.current.animator,
+			Object.assign(themeScrollContainerHandle.current, {
 				applyOverscrollEffect,
-				bounds: mutableRef.current.bounds,
 				calculateDistanceByWheel,
 				canScrollHorizontally,
 				canScrollVertically,
 				checkAndApplyOverscrollEffect,
 				getScrollBounds,
-				get isDragging () {
-					return mutableRef.current.isDragging;
-				},
-				set isDragging (val) {
-					mutableRef.current.isDragging = val;
-				},
-				get isScrollAnimationTargetAccumulated () {
-					return mutableRef.current.isScrollAnimationTargetAccumulated;
-				},
-				set isScrollAnimationTargetAccumulated (val) {
-					mutableRef.current.isScrollAnimationTargetAccumulated = val;
-				},
-				get lastInputType () {
-					return mutableRef.current.lastInputType;
-				},
-				set lastInputType (val) {
-					mutableRef.current.lastInputType = val;
-				},
-				get rtl () {
-					return rtl;
-				},
-				get scrollBounds () {
-					return getScrollBounds();
-				},
-				get scrollHeight () {
-					return mutableRef.current.bounds.scrollHeight;
-				},
-				get scrolling () {
-					return mutableRef.current.scrolling;
-				},
-				get scrollLeft () {
-					return mutableRef.current.scrollLeft;
-				},
 				scrollTo,
 				scrollToAccumulatedTarget,
-				get scrollToInfo () {
-					return mutableRef.current.scrollToInfo;
-				},
-				get scrollTop () {
-					return mutableRef.current.scrollTop;
-				},
 				setOverscrollStatus,
 				showScrollbarTrack,
 				start,
 				startHidingScrollbarTrack,
-				stop,
-				get wheelDirection () {
-					return mutableRef.current.wheelDirection;
-				},
-				set wheelDirection (val) {
-					mutableRef.current.wheelDirection = val;
-				}
+				stop
 			});
+			setScrollContainerHandle(themeScrollContainerHandle.current);
 		}
 	});
 
@@ -360,13 +375,14 @@ const useScrollBase = (props) => {
 
 		// `handleSize` in `ui/resolution.ResolutionDecorator` should be executed first.
 		setTimeout(() => {
-			if (propsHandleResizeWindow) {
-				propsHandleResizeWindow();
-			}
-			if (scrollMode === 'translate') {
-				scrollTo({position: {x: 0, y: 0}, animate: false});
-			} else {
-				scrollContentHandle.current.scrollToPosition(0, 0, 'instant');
+			const handleResizeResult = propsHandleResizeWindow?.();
+
+			if (handleResizeResult) {
+				if (scrollMode === 'translate') {
+					scrollTo({position: {x: 0, y: 0}, animate: false});
+				} else {
+					scrollContentHandle.current.scrollToPosition(0, 0, 'instant');
+				}
 			}
 
 			enqueueForceUpdate();
@@ -402,7 +418,7 @@ const useScrollBase = (props) => {
 		};
 	}, [direction, isHorizontalScrollbarVisible, isVerticalScrollbarVisible, rtl, scrollMode, spotlightContainerDisabled]); // eslint-disable-line react-hooks/exhaustive-deps
 
-	useEffect(() => {
+	useLayoutEffect(() => {
 		const
 			{hasDataSizeChanged} = scrollContentHandle.current,
 			{prevState, resizeRegistry, scrollToInfo} = mutableRef.current;