+++ packages/ui/CHANGELOG.md
@@ -2,6 +2,87 @@
 
 The following is a curated list of changes in the Enact ui module, newest changes on the top.
 
+## [4.5.3] - 2023-04-06
+
+No significant changes.
+## [4.7.0] - 2023-04-25
+
+### Added
+
+- `ui/Layout.Cell` prop `grow` to expand its size to the container
+
+### Fixed
+
+- `ui/ViewManager` to set index prop properly when reverseTransition prop is given
+
+## [4.6.2] - 2023-03-09
+
+No significant changes.
+
+## [4.6.1] - 2023-02-03
+
+### Added
+
+- `ui/Touchable` events `onPinch`, `onPinchStart`, `onPinchEnd`, and config `pinchConfig` to support pinch gesture
+
+### Deprecated
+
+- `ui/IconButton` to be removed in 5.0.0, use `ui/Button` instead
+- `ui/RadioDecorator` to be removed in 5.0.0
+- `ui/Scroller.ScrollerBase` to be removed in 5.0.0
+- `ui/SlotItem` to be removed in 5.0.0
+- `ui/ToggleItem` to be removed in 5.0.0
+- `ui/VirtualList.VirtualListBase` to be removed 5.0.0
+
+## [4.6.0] - 2022-12-05
+
+### Fixed
+
+- `ui/Marquee.MarqueeDecorator` to restart animation properly when `marqueeDelay` is 0
+
+## [4.0.12] - 2022-09-16
+
+### Fixed
+
+- `ui/Marquee.MarqueeDecorator` to have proper spacing for bidirectional text
+
+## [4.5.2] - 2022-08-17
+
+### Fixed
+
+- `ui/Marquee` to stop at the starting point after one cycle when scaled
+
+## [4.5.1] - 2022-08-03
+
+### Fixed
+
+- `ui/Marquee.MarqueeDecorator` to have proper spacing for bidirectional text
+- `ui/Marquee.MarqueeDecorator` to restart animation properly with React 18
+
+## [4.5.0] - 2022-07-19
+
+No significant changes.
+
+## [4.5.0-rc.2] - 2022-07-06
+
+### Fixed
+
+- `ui/Scroller` and `ui/VirtualList` to rerender property when `clientSize` is changed
+
+## [4.5.0-rc.1] - 2022-06-23
+
+No significant changes.
+
+## [4.5.0-beta.1] - 2022-05-31
+
+### Deprecated
+
+- `ui/GridListImageItem`, to be removed in 5.0.0. Use `ui/ImageItem` instead
+
+### Added
+
+- `ui/FloatingLayer` to add `detail` property containing `inputType` in `onDismiss` event payload
+
 ## [4.5.0-alpha.2] - 2022-05-09
 
 ### Deprecated
@@ -12,11 +93,19 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 - `ui/Marquee.MarqueeDecorator` config `css` to support customizing the marquee styles
 
+## [4.0.11] - 2022-04-25
+
+No significant changes.
+
 ## [4.5.0-alpha.1] - 2022-04-15
 
 ### Fixed
 
 - `ui/FloatingLayer` to stack popups always in the order in which they were opened
+
+## [4.0.10] - 2022-04-05
+
+No significant changes.
 
 ## [4.1.4] - 2022-03-24
 
@@ -37,12 +126,20 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 - Updated to use `forwardCustom` and add `type` when forwarding custom events
 
+## [3.2.7] - 2022-01-17
+
+No significant changes.
+
 ## [4.1.2] - 2021-12-22
 
 ### Fixed
 
 - `ui/Scroller` and `ui/VirtualList` to avoid stuttering of content on drag in mobile devices
 - `ui/Scroller` and `ui/VirtualList` to scroll correctly on Android platform for RTL locales
+
+## [4.0.9] - 2021-12-15
+
+No significant changes.
 
 ## [4.1.1] - 2021-11-30
 
@@ -339,6 +436,8 @@ No significant changes.
 
 ### Fixed
 
+- `ui/VirtualList.VirtualGridList` and `ui/VirtualList.VirtualList` to show items properly when reducing data size by updating `firstIndex` correctly
+- `ui/VirtualList.VirtualList` and `ui/VirtualList.VirtualGridList` not to suddenly jump when pressing directional keys after wheeling
 - `ui/VirtualList.VirtualList` and `ui/VirtualList.VirtualGridList` to show items properly when reducing data size
 
 ## [3.2.5] - 2019-11-14