+++ packages/webos/CHANGELOG.md
@@ -2,11 +2,64 @@
 
 The following is a curated list of changes in the Enact webos module, newest changes on the top.
 
+## [4.5.3] - 2023-04-06
+## [4.7.0] - 2023-04-25
+
+No significant changes.
+
+## [4.6.2] - 2023-03-09
+
+No significant changes.
+
+## [4.6.1] - 2023-02-03
+
+No significant changes.
+
+## [4.6.0] - 2022-12-05
+
+No significant changes.
+
+## [4.0.12] - 2022-09-16
+
+No significant changes.
+
+## [4.5.2] - 2022-08-17
+
+No significant changes.
+
+## [4.5.1] - 2022-08-03
+
+No significant changes.
+
+## [4.5.0] - 2022-07-19
+
+No significant changes.
+
+## [4.5.0-rc.2] - 2022-07-06
+
+No significant changes.
+
+## [4.5.0-rc.1] - 2022-06-23
+
+No significant changes.
+
+## [4.5.0-beta.1] - 2022-05-31
+
+No significant changes.
+
 ## [4.5.0-alpha.2] - 2022-05-09
 
 No significant changes.
 
+## [4.0.11] - 2022-04-25
+
+No significant changes.
+
 ## [4.5.0-alpha.1] - 2022-04-15
+
+No significant changes.
+
+## [4.0.10] - 2022-04-05
 
 No significant changes.
 
@@ -18,7 +71,15 @@ No significant changes.
 
 No significant changes.
 
+## [3.2.7] - 2022-01-17
+
+No significant changes.
+
 ## [4.1.2] - 2021-12-22
+
+No significant changes.
+
+## [4.0.9] - 2021-12-15
 
 No significant changes.